add turn alias and separate group and dataset
