+++ pyrit/attacks/multi_turn/tree_of_attacks.py
@@ -817,13 +817,11 @@ class _TreeOfAttacksNode:
         """
         # Configure for JSON response
         prompt_metadata: dict[str, str | int] = {"response_format": "json"}
-        seed_prompt_group = SeedPromptGroup(
-            prompts=[SeedPrompt(value=prompt_text, data_type="text", metadata=prompt_metadata)]
-        )
+        seed_prompts = [SeedPrompt(value=prompt_text, data_type="text", metadata=prompt_metadata)]
 
         # Send and get response
         response = await self._prompt_normalizer.send_prompt_async(
-            seed_prompt_group=seed_prompt_group,
+            seed_prompts=seed_prompts,
             conversation_id=self.adversarial_chat_conversation_id,
             target=self._adversarial_chat,
             labels=self._memory_labels,
