+++ pyrit/cli/__main__.py
@@ -10,7 +10,8 @@ from typing import List
 
 from pyrit.common import initialize_pyrit
 from pyrit.memory import CentralMemory
-from pyrit.models.seed_prompt import SeedPromptDataset, SeedPromptGroup
+from pyrit.models.seed_prompt_dataset import SeedPromptDataset
+from pyrit.models.seed_prompt_group import SeedPromptGroup
 
 from .scanner_config import ScannerConfig
