+++ pyrit/models/__init__.py
@@ -31,7 +31,9 @@ from pyrit.models.prompt_request_response import (
 from pyrit.models.prompt_response import PromptResponse
 from pyrit.models.question_answering import QuestionAnsweringDataset, QuestionAnsweringEntry, QuestionChoice
 from pyrit.models.score import Score, ScoreType, UnvalidatedScore
-from pyrit.models.seed_prompt import SeedPrompt, SeedPromptDataset, SeedPromptGroup
+from pyrit.models.seed_prompt import SeedPrompt
+from pyrit.models.seed_prompt_dataset import SeedPromptDataset
+from pyrit.models.seed_prompt_group import SeedPromptGroup
 from pyrit.models.storage_io import AzureBlobStorageIO, DiskStorageIO, StorageIO
 
 __all__ = [
@@ -72,6 +74,7 @@ __all__ = [
     "SeedPrompt",
     "SeedPromptDataset",
     "SeedPromptGroup",
+    "SeedPromptTurn",
     "StorageIO",
     "TextDataTypeSerializer",
     "UnvalidatedScore",
