+++ tests/integration/datasets/test_fetch_datasets.py
@@ -27,7 +27,7 @@ from pyrit.datasets import (
     fetch_wmdp_dataset,
     fetch_xstest_dataset,
 )
-from pyrit.models.seed_prompt import SeedPromptDataset
+from pyrit.models.seed_prompt_dataset import SeedPromptDataset
 
 
 @pytest.mark.parametrize(
