+++ tests/unit/test_normalizer_request.py
@@ -6,7 +6,8 @@ import uuid
 
 import pytest
 
-from pyrit.models.seed_prompt import SeedPrompt, SeedPromptGroup
+from pyrit.models.seed_prompt import SeedPrompt
+from pyrit.models.seed_prompt_group import SeedPromptGroup
 from pyrit.prompt_normalizer import NormalizerRequest
